+++ src/domain/api/utils/getIsError.ts
@@ -1,0 +1,7 @@
+import { GraphQLError } from 'graphql';
+
+function getIsError(graphQLError: GraphQLError, errorType: string): boolean {
+  return graphQLError.extensions?.code === errorType;
+}
+
+export default getIsError;